requirements.txt
langchain==0.1.12
langchain-openai==0.0.8
python-dotenv==1.0.1
chromadb==0.4.24
tiktoken==0.6.0
requests==2.31.0
sentence-transformers==2.5.1 